requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-jose[cryptography]>=3.3.0
httpx>=0.25.0
pyyaml>=6.0.1
websockets>=12.0
